Fix: (Editor) textfield pre-filled on IFrame load, not immediately
Previously, in edit_newsletter.html, with a specified 'nid' to preload,
because the corresponding AJAX request is async, if the contents
of the Editor JFrame were not yet loaded, the textarea would not
have been pre-filled.
This commit fixes this, filling the textarea only after the IFrame
has loaded (making use of $#on('load', `...`)).